+++ apps/web/src/pages/Profile/ExperienceFormPage/ExperienceForm.stories.tsx
@@ -3,8 +3,11 @@ import type { Meta, Story } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 import { getStaticSkills } from "@gc-digital-talent/fake-data";
 import type { Skill } from "~/api/generated";
+import type {
+  ExperienceDetailsSubmissionData,
+  ExperienceType,
+} from "~/types/experience";
 import { ExperienceForm } from "./ExperienceFormPage";
-import type { ExperienceDetailsSubmissionData, ExperienceType } from "./types";
 
 const skillData = getStaticSkills();
